Fix Cart tests

+++ src/test/java/com/leroy/magmobile/api/data/sales/cart_estimate/cart/CartDiscountData.java
@@ -8,7 +8,7 @@ import java.time.LocalDateTime;
 @Data
 public class CartDiscountData {
     private String type;
-    private Integer typeValue;
+    private Double typeValue;
     private String actor;
     @JsonFormat(pattern = "yyyy-MM-dd'T'HH:mm:ss.SSS'Z'")
     private LocalDateTime updated;
